+++ examples/derive_value_impl.rs
@@ -25,7 +25,7 @@ fn main() -> Result<(), Error> {
 
     let read_txn = db.begin_read()?;
     let table = read_txn.open_table(TABLE)?;
-    assert_eq!(table.get(&key)?.unwrap().value(), 0);
+    assert_eq!(table.get_pinned(&key)?.unwrap().value(), 0);
 
     Ok(())
 }